+++ rust/dbn/src/encode/json.rs
@@ -51,8 +51,6 @@ where
             self.use_pretty_ts,
         );
         self.writer.write_all(json.as_bytes())?;
-        // newline at EOF
-        self.writer.write_all(b"\n")?;
         self.writer.flush()?;
         Ok(())
     }
@@ -85,10 +83,6 @@ where
             Err(e) => {
                 Err(anyhow::Error::new(e).context(format!("Failed to serialize {record:#?}")))
             }
-        }?;
-        match self.writer.write_all(b"\n") {
-            Err(e) if e.kind() == io::ErrorKind::BrokenPipe => return Ok(true),
-            r => r,
         }?;
         Ok(false)
     }
@@ -151,6 +145,7 @@ pub(crate) mod serialize {
             let mut writer = JsonObjectWriter::new(&mut res);
             to_json_with_writer(obj, &mut writer, use_pretty_px, use_pretty_ts);
         }
+        res.push('\n');
         res
     }
 
@@ -514,7 +509,7 @@ mod tests {
         String::from_utf8(buffer).expect("valid UTF-8")
     }
 
-    const HEADER_JSON: &str =
+    pub const HEADER_JSON: &str =
         r#""hd":{"rtype":4,"publisher_id":1,"instrument_id":323,"ts_event":"1658441851000000000"}"#;
     const BID_ASK_JSON: &str = r#"{"bid_px":"372000.000000000","ask_px":"372500.000000000","bid_sz":10,"ask_sz":5,"bid_ct":5,"ask_ct":2}"#;
 
@@ -925,5 +920,222 @@ mod tests {
             r#"{"hd":{"rtype":21,"publisher_id":0,"instrument_id":0,"ts_event":null},"err":"\"A test"}
 "#
         );
+    }
+}
+
+#[cfg(feature = "async")]
+pub use r#async::Encoder as AsyncEncoder;
+
+#[cfg(feature = "async")]
+mod r#async {
+    use tokio::io;
+
+    use crate::{
+        encode::DbnEncodable, record_ref::RecordRef, rtype_ts_out_async_dispatch, Metadata,
+    };
+
+    /// Type for encoding files and streams of DBN records in newline-delimited JSON (ndjson).
+    pub struct Encoder<W>
+    where
+        W: io::AsyncWriteExt + Unpin,
+    {
+        writer: W,
+        should_pretty_print: bool,
+        use_pretty_px: bool,
+        use_pretty_ts: bool,
+    }
+
+    impl<W> Encoder<W>
+    where
+        W: io::AsyncWriteExt + Unpin,
+    {
+        /// Creates a new instance of [`Encoder`]. If `should_pretty_print` is `true`,
+        /// each JSON object will be nicely formatted and indented, instead of the default
+        /// compact output with no whitespace between key-value pairs.
+        pub fn new(
+            writer: W,
+            should_pretty_print: bool,
+            use_pretty_px: bool,
+            use_pretty_ts: bool,
+        ) -> Self {
+            Self {
+                writer,
+                should_pretty_print,
+                use_pretty_px,
+                use_pretty_ts,
+            }
+        }
+
+        /// Encodes `metadata` into JSON.
+        ///
+        /// # Errors
+        /// This function returns an error if there's an error writing to `writer`.
+        pub async fn encode_metadata(&mut self, metadata: &Metadata) -> anyhow::Result<()> {
+            let json = super::to_json_string(
+                metadata,
+                self.should_pretty_print,
+                self.use_pretty_px,
+                self.use_pretty_ts,
+            );
+            self.writer.write_all(json.as_bytes()).await?;
+            self.writer.flush().await?;
+            Ok(())
+        }
+
+        /// Returns a reference to the underlying writer.
+        pub fn get_ref(&self) -> &W {
+            &self.writer
+        }
+
+        /// Returns a mutable reference to the underlying writer.
+        pub fn get_mut(&mut self) -> &mut W {
+            &mut self.writer
+        }
+
+        /// Encode a single DBN record of type `R`.
+        ///
+        /// Returns `true`if the pipe was closed.
+        ///
+        /// # Errors
+        /// This function returns an error if it's unable to write to the underlying
+        /// writer.
+        pub async fn encode_record<R: DbnEncodable>(&mut self, record: &R) -> anyhow::Result<bool> {
+            let json = super::to_json_string(
+                record,
+                self.should_pretty_print,
+                self.use_pretty_px,
+                self.use_pretty_ts,
+            );
+            match self.writer.write_all(json.as_bytes()).await {
+                Ok(_) => Ok(()),
+                Err(e) if matches!(e.kind(), io::ErrorKind::BrokenPipe) => return Ok(true),
+                Err(e) => {
+                    Err(anyhow::Error::new(e).context(format!("Failed to serialize {record:#?}")))
+                }
+            }?;
+            Ok(false)
+        }
+
+        /// Encodes a single DBN record.
+        ///
+        /// Returns `true`if the pipe was closed.
+        ///
+        /// # Safety
+        /// `ts_out` must be `false` if `record` does not have an appended `ts_out
+        ///
+        /// # Errors
+        /// This function returns an error if it's unable to write to the underlying writer
+        /// or there's a serialization error.
+        pub async unsafe fn encode_record_ref(
+            &mut self,
+            record_ref: RecordRef<'_>,
+            ts_out: bool,
+        ) -> anyhow::Result<bool> {
+            rtype_ts_out_async_dispatch!(record_ref, ts_out, |rec| async move {
+                self.encode_record(rec).await
+            })?
+        }
+    }
+
+    #[cfg(test)]
+    mod tests {
+        use std::ffi::c_char;
+
+        use tokio::io::{AsyncWriteExt, BufWriter};
+
+        use crate::{
+            encode::test_data::RECORD_HEADER,
+            enums::rtype,
+            record::{HasRType, MboMsg, RecordHeader},
+        };
+
+        use super::*;
+
+        async fn write_to_json_string<R>(
+            record: &R,
+            should_pretty_print: bool,
+            use_pretty_px: bool,
+            use_pretty_ts: bool,
+        ) -> String
+        where
+            R: DbnEncodable,
+        {
+            let mut buffer = Vec::new();
+            let mut writer = BufWriter::new(&mut buffer);
+            Encoder::new(
+                &mut writer,
+                should_pretty_print,
+                use_pretty_px,
+                use_pretty_ts,
+            )
+            .encode_record(record)
+            .await
+            .unwrap();
+            writer.flush().await.unwrap();
+            String::from_utf8(buffer).expect("valid UTF-8")
+        }
+
+        async fn write_ref_to_json_string(
+            record: RecordRef<'_>,
+            should_pretty_print: bool,
+            use_pretty_px: bool,
+            use_pretty_ts: bool,
+        ) -> String {
+            let mut buffer = Vec::new();
+            let mut writer = BufWriter::new(&mut buffer);
+            unsafe {
+                Encoder::new(
+                    &mut writer,
+                    should_pretty_print,
+                    use_pretty_px,
+                    use_pretty_ts,
+                )
+                .encode_record_ref(record, false)
+            }
+            .await
+            .unwrap();
+            writer.flush().await.unwrap();
+            String::from_utf8(buffer).expect("valid UTF-8")
+        }
+
+        #[tokio::test]
+        async fn test_mbo_write_json() {
+            let record = MboMsg {
+                hd: RecordHeader::new::<MboMsg>(
+                    rtype::MBO,
+                    RECORD_HEADER.publisher_id,
+                    RECORD_HEADER.instrument_id,
+                    RECORD_HEADER.ts_event,
+                ),
+                order_id: 16,
+                price: 5500,
+                size: 3,
+                flags: 128,
+                channel_id: 14,
+                action: 'R' as c_char,
+                side: 'N' as c_char,
+                ts_recv: 1658441891000000000,
+                ts_in_delta: 22_000,
+                sequence: 1_002_375,
+            };
+            let res = write_to_json_string(&record, false, true, false).await;
+            let ref_res = write_ref_to_json_string(
+                unsafe { RecordRef::unchecked_from_header(record.header()) },
+                false,
+                true,
+                false,
+            )
+            .await;
+
+            assert_eq!(res, ref_res);
+            assert_eq!(
+                ref_res,
+                format!(
+                    "{{{},{}}}\n",
+                    r#""hd":{"rtype":160,"publisher_id":1,"instrument_id":323,"ts_event":"1658441851000000000"}"#,
+                    r#""order_id":"16","price":"0.000005500","size":3,"flags":128,"channel_id":14,"action":"R","side":"N","ts_recv":"1658441891000000000","ts_in_delta":22000,"sequence":1002375"#
+                )
+            );
+        }
     }
 }